localized a bug and added readme
1. bug was due to different java used on cluster
2. added a readme

+++ pipeline_1KGP/runRASQUAL/step2_permute_VCF.R
@@ -21,25 +21,35 @@ geneid = unique(genes[,11])
 to.rm = grep(":", geneid)
 if(length(to.rm)>0)geneid=geneid[-to.rm]
 
-res = data.frame(matrix(NA, nrow=length(geneid), ncol=3))
-win = 2e5
-for(i in 1:length(geneid)){
-  #m = match(geneid[i], genes[,11])
-  gnm = which(geneid[i]==genes[,11])
-  chri = genes[gnm[1],2]
-  system(sprintf("grep '%s ' tmpsnps.txt > chrisnps.txt", chri))  
-  chrsnps = read.table("chrisnps.txt")
-  posst = min(genes[gnm,5])
-  posen = max(genes[gnm,6])
-  kp = which(((posst-win)<=chrsnps[,2]) & ((posen+win)>chrsnps[,2]))
-  res[i, 1] = chri
-  res[i, 2] = geneid[i]
-  res[i, 3] = length(kp)
-  message(i)
+#
+#a.
+#count number of SNPs within a window - will be used as input for RASQUAL
+#
+snp.cnt = sprintf("%s/number_of_snps_per_gene_win%s.csv", info.dir, win)
+if(!file.exists(snp.cnt)){
+  res = data.frame(matrix(NA, nrow=length(geneid), ncol=3))
+  win = 2e5
+  for(i in 1:length(geneid)){
+    #m = match(geneid[i], genes[,11])
+    gnm = which(geneid[i]==genes[,11])
+    chri = genes[gnm[1],2]
+    system(sprintf("grep '%s ' tmpsnps.txt > chrisnps.txt", chri))  
+    chrsnps = read.table("chrisnps.txt")
+    posst = min(genes[gnm,5])
+    posen = max(genes[gnm,6])
+    kp = which(((posst-win)<=chrsnps[,2]) & ((posen+win)>chrsnps[,2]))
+    res[i, 1] = chri
+    res[i, 2] = geneid[i]
+    res[i, 3] = length(kp)
+    message(i)
+  }
+  write.csv(res, sprintf("%s/number_of_snps_per_gene_win%s.csv", info.dir, win), quote=F, row.names=F) 
 }
-write.csv(res, sprintf("%s/number_of_snps_per_gene_win%s.csv", info.dir, win), quote=F, row.names=F) 
 
-#add vcf counts
+#
+#b.
+#add snp level allele-specific counts to vcf
+#
 vcf.hd = read.table(snp, as.is=T, nrow=4, comment.char=",", sep=",")
 vcf.in = read.table(snp, as.is=T)
 
@@ -49,7 +59,7 @@ count.ref = function(vec){mean(unlist(strsplit(vec, split="\\|"))==1)}
 af = apply(vcf.in[,-(1:9)], 1, count.ref)
 vcf.in[,8] = af
 
-#if too many reads of wrong bases - may skip this count
+#if too many reads come from wrong bases - may skip this count
 #also add a permutation step for VCF
 frac = 0.1
 samples = unlist(strsplit(vcf.hd[4,1], split="\t"))[-(1:9)]
@@ -76,8 +86,10 @@ system(sprintf("bgzip -f %s", out.vcf))
 out.vcf = sprintf("%s.gz", out.vcf)
 system(sprintf("tabix %s", out.vcf))
 
-
+#
+#c
 #create other required files if they don't exist yet
+#
 totc = sprintf("%s/Tcnt_ex.dat", cnt.dir)
 xcnt = sprintf("%s/samples.dat", cnt.dir)
 totc = read.table(totc, as.is=T)
@@ -87,6 +99,9 @@ totb = sprintf("%s/Tcnt_ex.bin", cnt.dir)
 kbin = sprintf("%s/K_ex.bin", cnt.dir)
 xbin = sprintf("%s/X_ex.bin", cnt.dir)
 
+#note, transposing matrices is important to so that vector of values contains
+#gene1, gene2, ... genen information consequtively
+#also, it is important to ensure that data is stored as double
 if(!file.exists(totb)){
   con = file(totb, "wb")
   writeBin(as.double(t(totc)), con=con, double())
